chore: renamed and shuffled some files around
