+++ test/functionsMetaData.spec.ts
@@ -2,7 +2,7 @@ import { readJSON } from 'fs-extra'
 import mock from 'mock-fs'
 import { join } from 'pathe'
 import { NEXT_PLUGIN_NAME } from '../packages/runtime/src/constants'
-import { writeFunctionConfiguration } from '../packages/runtime/src/helpers/utilities/functions'
+import { writeFunctionConfiguration } from '../packages/runtime/src/helpers/functionsMetaData'
 
 describe('writeFunctionConfiguration', () => {
   afterEach(() => {
